SIPUSH bytecode is not used for short integer constant #33 Also streamlined the handling of line numbers.

+++ janino/src/main/java/org/codehaus/janino/Java.java
@@ -3233,7 +3233,7 @@ class Java {
             if (x != 0) {
                 CodeContext ca = this.getCodeContext();
                 ca.pushInserter(this);
-                ca.makeSpace((short) -1, 4 - x);
+                ca.makeSpace(-1, 4 - x);
                 ca.popInserter();
             }
         }
